Split users and rooms routes
