+++ client/admin/users/UsersRoute.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import { usePermission } from '../../contexts/AuthorizationContext';
+import NotAuthorizedPage from '../NotAuthorizedPage';
+import UsersPage from './UsersPage';
+
+function UsersRoute() {
+	const canViewUserAdministration = usePermission('view-user-administration');
+
+	if (!canViewUserAdministration) {
+		return <NotAuthorizedPage />;
+	}
+
+	return <UsersPage />;
+}
+
+export default UsersRoute;
